more restructuring. gui (including state) in one folder
