+++ src/gui/message.rs
@@ -1,0 +1,9 @@
+// Tim Lobner
+
+#[derive(Debug, Clone)]
+pub enum Message{
+    ChooseRunningOrderInput,
+    OnRunningOrderInputChanged(String),
+    CreateCompleteRunningOrder,
+    CreatePersonalRunningOrder,
+}
